+++ src/Components/Instructions/Instructions.js
@@ -21,7 +21,6 @@ export default class Instructions extends Component {
             <img
               className='logo'
               height='80px'
-              className='center'
               alt='app logo'
               src={logo}></img>
           </header>
