+++ src/Components/Login/LoginFormContainer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import LoginForm from './LoginForm';
-import { login } from '../../actions/user';
+import { login } from '../../actions/auth';
 import { connect } from 'react-redux';
 
 class LoginFormContainer extends React.Component {
@@ -10,6 +10,8 @@ class LoginFormContainer extends React.Component {
     event.preventDefault();
     this.props.login(this.state.email, this.state.code);
     this.props.history.push('/instructions');
+    //  this old code, will redirect event though the onSubmit is wrong (even though either email or code is not correct, fix this to only work when email + code is correct
+    /// use redirect instead in render()???
   };
 
   onChange = event => {
@@ -28,6 +30,12 @@ class LoginFormContainer extends React.Component {
     );
   }
 }
+
+// const mapStateToProps = state => {
+//   return {
+//     token: state.auth
+//   };
+// };
 
 export default connect(
   null,
