+++ src/reducers/index.js
@@ -1,10 +1,10 @@
 import { combineReducers } from 'redux';
-import user from '../reducers/user';
+import auth from '../reducers/auth';
 import question from '../reducers/question';
 import answer from '../reducers/answer';
 
 export default combineReducers({
-  user,
+  auth,
   question,
   answer,
 });
